+++ database/seeders/TutorTutorialSeeder.php
@@ -1,0 +1,92 @@
+<?php
+
+namespace Database\Seeders;
+
+use Illuminate\Database\Seeder;
+
+class TutorTutorialSeeder extends Seeder
+{
+    /**
+     * Run the database seeds.
+     *
+     * @return void
+     */
+    public function run()
+    {
+        \Illuminate\Support\Facades\DB::table('tutor_tutorial')
+            ->insertOrIgnore([
+                [
+                    'tutor_id' => 'pcw1',
+                    'tutorial_id' => 1,
+                    'created_at' => date('Y-m-d H:i:s'),
+                    'updated_at' => date('Y-m-d H:i:s'),
+                ],
+                [
+                    'tutor_id' => 'flk1',
+                    'tutorial_id' => 2,
+                    'created_at' => date('Y-m-d H:i:s'),
+                    'updated_at' => date('Y-m-d H:i:s'),
+                ],
+                [
+                    'tutor_id' => 'jel1',
+                    'tutorial_id' => 2,
+                    'created_at' => date('Y-m-d H:i:s'),
+                    'updated_at' => date('Y-m-d H:i:s'),
+                ],
+                [
+                    'tutor_id' => 'rao1',
+                    'tutorial_id' => 2,
+                    'created_at' => date('Y-m-d H:i:s'),
+                    'updated_at' => date('Y-m-d H:i:s'),
+                ],
+                [
+                    'tutor_id' => 'bb11',
+                    'tutorial_id' => 3,
+                    'created_at' => date('Y-m-d H:i:s'),
+                    'updated_at' => date('Y-m-d H:i:s'),
+                ],
+                [
+                    'tutor_id' => 'aj34',
+                    'tutorial_id' => 4,
+                    'created_at' => date('Y-m-d H:i:s'),
+                    'updated_at' => date('Y-m-d H:i:s'),
+                ],
+                [
+                    'tutor_id' => 'bb11',
+                    'tutorial_id' => 5,
+                    'created_at' => date('Y-m-d H:i:s'),
+                    'updated_at' => date('Y-m-d H:i:s'),
+                ],
+                [
+                    'tutor_id' => 'kch1',
+                    'tutorial_id' => 6,
+                    'created_at' => date('Y-m-d H:i:s'),
+                    'updated_at' => date('Y-m-d H:i:s'),
+                ],
+                [
+                    'tutor_id' => 'gdm1',
+                    'tutorial_id' => 7,
+                    'created_at' => date('Y-m-d H:i:s'),
+                    'updated_at' => date('Y-m-d H:i:s'),
+                ],
+                [
+                    'tutor_id' => 'cib1',
+                    'tutorial_id' => 8,
+                    'created_at' => date('Y-m-d H:i:s'),
+                    'updated_at' => date('Y-m-d H:i:s'),
+                ],
+                [
+                    'tutor_id' => 'jjc1',
+                    'tutorial_id' => 8,
+                    'created_at' => date('Y-m-d H:i:s'),
+                    'updated_at' => date('Y-m-d H:i:s'),
+                ],
+                [
+                    'tutor_id' => 'oaa2',
+                    'tutorial_id' => 9,
+                    'created_at' => date('Y-m-d H:i:s'),
+                    'updated_at' => date('Y-m-d H:i:s'),
+                ],
+            ]);
+    }
+}
