+++ database/seeders/TutorialSeeder.php
@@ -1,8 +1,10 @@
 <?php
 
+namespace Database\Seeders;
+
 use Illuminate\Database\Seeder;
 
-class UserSeeder extends Seeder
+class TutorialSeeder extends Seeder
 {
     /**
      * Run the database seeds.
@@ -11,85 +13,77 @@ class UserSeeder extends Seeder
      */
     public function run()
     {
-        \App\User::factory()->count(3)->create();
-
-        \Illuminate\Support\Facades\DB::table('module_user')
+        \Illuminate\Support\Facades\DB::table('tutorials')
             ->insertOrIgnore([
                 [
-                    'module_id' => 'COSE50637',
-                    'user_id' => 1,
-                    'created_at' => date('Y-m-d H:i:s'),
-                    'updated_at' => date('Y-m-d H:i:s'),
-                ],
-                [
-                    'module_id' => 'COSE50586',
-                    'user_id' => 1,
+                    'module_id' => 'COMP50016',
+                    'time_start' => '2020-01-20 10:00:00',
+                    'time_end' => '2020-01-20 13:00:00',
+                    'room' => 'S507',
                     'created_at' => date('Y-m-d H:i:s'),
                     'updated_at' => date('Y-m-d H:i:s'),
                 ],
                 [
                     'module_id' => 'COMP50017',
-                    'user_id' => 1,
-                    'created_at' => date('Y-m-d H:i:s'),
-                    'updated_at' => date('Y-m-d H:i:s'),
-                ],
-                [
-                    'module_id' => 'COMP50016',
-                    'user_id' => 1,
-                    'created_at' => date('Y-m-d H:i:s'),
-                    'updated_at' => date('Y-m-d H:i:s'),
-                ],
-                [
-                    'module_id' => 'COSE50637',
-                    'user_id' => 2,
+                    'time_start' => '2020-01-24 14:00:00',
+                    'time_end' => '2020-01-24 17:00:00',
+                    'room' => 'S503',
                     'created_at' => date('Y-m-d H:i:s'),
                     'updated_at' => date('Y-m-d H:i:s'),
                 ],
                 [
                     'module_id' => 'COSE50586',
-                    'user_id' => 2,
+                    'time_start' => '2020-01-24 11:00:00',
+                    'time_end' => '2020-01-24 13:00:00',
+                    'room' => 'S509',
+                    'created_at' => date('Y-m-d H:i:s'),
+                    'updated_at' => date('Y-m-d H:i:s'),
+                ],
+                [
+                    'module_id' => 'COSE50586',
+                    'time_start' => '2020-01-23 15:00:00',
+                    'time_end' => '2020-01-23 17:00:00',
+                    'room' => 'S503',
+                    'created_at' => date('Y-m-d H:i:s'),
+                    'updated_at' => date('Y-m-d H:i:s'),
+                ],
+                [
+                    'module_id' => 'COSE50637',
+                    'time_start' => '2020-01-24 09:00:00',
+                    'time_end' => '2020-01-24 10:00:00',
+                    'room' => 'S507',
+                    'created_at' => date('Y-m-d H:i:s'),
+                    'updated_at' => date('Y-m-d H:i:s'),
+                ],
+                [
+                    'module_id' => 'COSE50637',
+                    'time_start' => '2020-01-24 10:00:00',
+                    'time_end' => '2020-01-24 11:00:00',
+                    'room' => 'S507',
                     'created_at' => date('Y-m-d H:i:s'),
                     'updated_at' => date('Y-m-d H:i:s'),
                 ],
                 [
                     'module_id' => 'COSE50582',
-                    'user_id' => 2,
-                    'created_at' => date('Y-m-d H:i:s'),
-                    'updated_at' => date('Y-m-d H:i:s'),
-                ],
-                [
-                    'module_id' => 'COMP50016',
-                    'user_id' => 2,
-                    'created_at' => date('Y-m-d H:i:s'),
-                    'updated_at' => date('Y-m-d H:i:s'),
-                ],
-                [
-                    'module_id' => 'COSE50637',
-                    'user_id' => 3,
-                    'created_at' => date('Y-m-d H:i:s'),
-                    'updated_at' => date('Y-m-d H:i:s'),
-                ],
-                [
-                    'module_id' => 'COSE50586',
-                    'user_id' => 3,
-                    'created_at' => date('Y-m-d H:i:s'),
-                    'updated_at' => date('Y-m-d H:i:s'),
-                ],
-                [
-                    'module_id' => 'COMP50016',
-                    'user_id' => 3,
-                    'created_at' => date('Y-m-d H:i:s'),
-                    'updated_at' => date('Y-m-d H:i:s'),
-                ],
-                [
-                    'module_id' => 'COIS51092',
-                    'user_id' => 3,
+                    'time_start' => '2020-01-21 14:00:00',
+                    'time_end' => '2020-01-21 17:00:00',
+                    'room' => 'S507',
                     'created_at' => date('Y-m-d H:i:s'),
                     'updated_at' => date('Y-m-d H:i:s'),
                 ],
                 [
                     'module_id' => 'COIS51091',
-                    'user_id' => 3,
+                    'time_start' => '2019-09-25 09:00:00',
+                    'time_end' => '2020-09-25 12:00:00',
+                    'room' => 'S205',
+                    'created_at' => date('Y-m-d H:i:s'),
+                    'updated_at' => date('Y-m-d H:i:s'),
+                ],
+                [
+                    'module_id' => 'COIS51092',
+                    'time_start' => '2019-09-23 13:00:00',
+                    'time_end' => '2020-09-23 15:00:00',
+                    'room' => 'S501',
                     'created_at' => date('Y-m-d H:i:s'),
                     'updated_at' => date('Y-m-d H:i:s'),
                 ],
