+++ database/seeders/UserSeeder.php
@@ -1,8 +1,10 @@
 <?php
 
+namespace Database\Seeders;
+
 use Illuminate\Database\Seeder;
 
-class TutorTutorialSeeder extends Seeder
+class UserSeeder extends Seeder
 {
     /**
      * Run the database seeds.
@@ -11,77 +13,85 @@ class TutorTutorialSeeder extends Seeder
      */
     public function run()
     {
-        \Illuminate\Support\Facades\DB::table('tutor_tutorial')
+        \App\User::factory()->count(3)->create();
+
+        \Illuminate\Support\Facades\DB::table('module_user')
             ->insertOrIgnore([
                 [
-                    'tutor_id' => 'pcw1',
-                    'tutorial_id' => 1,
+                    'module_id' => 'COSE50637',
+                    'user_id' => 1,
                     'created_at' => date('Y-m-d H:i:s'),
                     'updated_at' => date('Y-m-d H:i:s'),
                 ],
                 [
-                    'tutor_id' => 'flk1',
-                    'tutorial_id' => 2,
+                    'module_id' => 'COSE50586',
+                    'user_id' => 1,
                     'created_at' => date('Y-m-d H:i:s'),
                     'updated_at' => date('Y-m-d H:i:s'),
                 ],
                 [
-                    'tutor_id' => 'jel1',
-                    'tutorial_id' => 2,
+                    'module_id' => 'COMP50017',
+                    'user_id' => 1,
                     'created_at' => date('Y-m-d H:i:s'),
                     'updated_at' => date('Y-m-d H:i:s'),
                 ],
                 [
-                    'tutor_id' => 'rao1',
-                    'tutorial_id' => 2,
+                    'module_id' => 'COMP50016',
+                    'user_id' => 1,
                     'created_at' => date('Y-m-d H:i:s'),
                     'updated_at' => date('Y-m-d H:i:s'),
                 ],
                 [
-                    'tutor_id' => 'bb11',
-                    'tutorial_id' => 3,
+                    'module_id' => 'COSE50637',
+                    'user_id' => 2,
                     'created_at' => date('Y-m-d H:i:s'),
                     'updated_at' => date('Y-m-d H:i:s'),
                 ],
                 [
-                    'tutor_id' => 'aj34',
-                    'tutorial_id' => 4,
+                    'module_id' => 'COSE50586',
+                    'user_id' => 2,
                     'created_at' => date('Y-m-d H:i:s'),
                     'updated_at' => date('Y-m-d H:i:s'),
                 ],
                 [
-                    'tutor_id' => 'bb11',
-                    'tutorial_id' => 5,
+                    'module_id' => 'COSE50582',
+                    'user_id' => 2,
                     'created_at' => date('Y-m-d H:i:s'),
                     'updated_at' => date('Y-m-d H:i:s'),
                 ],
                 [
-                    'tutor_id' => 'kch1',
-                    'tutorial_id' => 6,
+                    'module_id' => 'COMP50016',
+                    'user_id' => 2,
                     'created_at' => date('Y-m-d H:i:s'),
                     'updated_at' => date('Y-m-d H:i:s'),
                 ],
                 [
-                    'tutor_id' => 'gdm1',
-                    'tutorial_id' => 7,
+                    'module_id' => 'COSE50637',
+                    'user_id' => 3,
                     'created_at' => date('Y-m-d H:i:s'),
                     'updated_at' => date('Y-m-d H:i:s'),
                 ],
                 [
-                    'tutor_id' => 'cib1',
-                    'tutorial_id' => 8,
+                    'module_id' => 'COSE50586',
+                    'user_id' => 3,
                     'created_at' => date('Y-m-d H:i:s'),
                     'updated_at' => date('Y-m-d H:i:s'),
                 ],
                 [
-                    'tutor_id' => 'jjc1',
-                    'tutorial_id' => 8,
+                    'module_id' => 'COMP50016',
+                    'user_id' => 3,
                     'created_at' => date('Y-m-d H:i:s'),
                     'updated_at' => date('Y-m-d H:i:s'),
                 ],
                 [
-                    'tutor_id' => 'oaa2',
-                    'tutorial_id' => 9,
+                    'module_id' => 'COIS51092',
+                    'user_id' => 3,
+                    'created_at' => date('Y-m-d H:i:s'),
+                    'updated_at' => date('Y-m-d H:i:s'),
+                ],
+                [
+                    'module_id' => 'COIS51091',
+                    'user_id' => 3,
                     'created_at' => date('Y-m-d H:i:s'),
                     'updated_at' => date('Y-m-d H:i:s'),
                 ],
